requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
python-dotenv==0.19.0
gunicorn==20.1.0
Werkzeug==2.0.1
SQLAlchemy==1.4.46
psycopg2-binary==2.8.6 